Add validtions

+++ src/main/java/com/ecommerce/book/controller/CategoryController.java
@@ -4,6 +4,7 @@ import com.ecommerce.book.dto.CreateCategoryDTO;
 import com.ecommerce.book.dto.ResponseCategoryDTO;
 import com.ecommerce.book.dto.UpdateCategoryDTO;
 import com.ecommerce.book.service.CategoryService;
+import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -21,20 +22,27 @@ public class CategoryController {
     // Removed CategoryMapper injection as it's already handled by CategoryService
 
     @PostMapping("/addcategory")
-    public ResponseEntity<ResponseCategoryDTO> addCategory(@RequestBody CreateCategoryDTO category) {
+    public ResponseEntity<ResponseCategoryDTO> addCategory(
+            @RequestBody @Valid CreateCategoryDTO category
+    ) {
         ResponseCategoryDTO categoryDTO = categoryService.addCategory(category);
         return ResponseEntity.ok(categoryDTO);
     }
 
     @PutMapping("/updater/{id}")
-    public ResponseEntity<ResponseCategoryDTO> changeCategory(@PathVariable int id, @RequestBody UpdateCategoryDTO category) {
+    public ResponseEntity<ResponseCategoryDTO> changeCategory(
+            @PathVariable int id, @RequestBody @Valid UpdateCategoryDTO category
+    ) {
         ResponseCategoryDTO categoryDTO = categoryService.updateCategory(id, category);
         return ResponseEntity.ok(categoryDTO);
     }
 
     @GetMapping("/category/{id}")
-    public ResponseEntity<Optional<ResponseCategoryDTO>> getCategory(@PathVariable int id) {
+    public ResponseEntity<Optional<ResponseCategoryDTO>> getCategory(@PathVariable  int id) {
         Optional<ResponseCategoryDTO> categoryDTO = Optional.ofNullable(categoryService.getCategoryById(id));
+        if (categoryDTO.isEmpty() ) {
+            return ResponseEntity.notFound().build();
+        }
         return ResponseEntity.ok(categoryDTO);
     }
 
@@ -45,7 +53,11 @@ public class CategoryController {
     }
 
     @DeleteMapping("/deletecategory/{id}")
-    public void deleteCategory(@PathVariable int id) {
+    public ResponseEntity<Void> deleteCategory(@PathVariable int id) {
+        if (id <=0){
+            return ResponseEntity.badRequest().build();
+        }
         categoryService.deleteCategory(id);
+        return ResponseEntity.ok().build();
     }
 }
